// frontend/src/api/generated/models/adjustmentDto.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * MH
 * OpenAPI spec version: 1.0.0
 */
import type { AdjustmentDtoType } from './adjustmentDtoType';

export interface AdjustmentDto {
  id: number;
  type: AdjustmentDtoType;
  reason: string;
  quantity: number;
  /** @nullable */
  occurredAt?: string | null;
  createdAt: string;
  /** @nullable */
  importationId?: number | null;
  /** @nullable */
  importationProductId?: number | null;
  /** @nullable */
  orderItemId?: number | null;
  /** @nullable */
  orderId?: string | null;
  /** @nullable */
  sourceTitle?: string | null;
  /** @nullable */
  sourceVariantName?: string | null;
  /** @nullable */
  sourceSku?: string | null;
  /** @nullable */
  targetTitle?: string | null;
  /** @nullable */
  targetVariantName?: string | null;
  /** @nullable */
  targetSku?: string | null;
}