// frontend/src/api/generated/models/adjustmentDtoType.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * MH
 * OpenAPI spec version: 1.0.0
 */

export type AdjustmentDtoType = typeof AdjustmentDtoType[keyof typeof AdjustmentDtoType];


export const AdjustmentDtoType = {
  destruction: 'destruction',
  mutation: 'mutation',
  swap: 'swap',
} as const;